+++ crates/ruma-signatures/src/lib.rs
@@ -51,14 +51,14 @@
 pub use ruma_common::{IdParseError, SigningKeyAlgorithm};
 
 pub use self::{
-    error::{Error, JsonError, ParseError, VerificationError},
+    error::{JsonError, VerificationError},
     functions::{
         canonical_json, content_hash, hash_and_sign_event, reference_hash, sign_json,
         verify_canonical_json_bytes, verify_event, verify_json,
     },
-    keys::{Ed25519KeyPair, KeyPair, PublicKeyMap, PublicKeySet},
+    keys::{Ed25519KeyPair, Ed25519KeyPairParseError, KeyPair, PublicKeyMap, PublicKeySet},
     signatures::Signature,
-    verification::Verified,
+    verification::{Ed25519VerificationError, Verified},
 };
 
 mod error;